examples/fashion_content_agent/requirements.txt
# Document Storage Dependencies
openai>=1.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
Pillow>=10.0.0
requests>=2.31.0
python-docx>=0.8.11
markdown>=3.0.0
airtable-python-wrapper>=0.15.0
streamlit>=1.32.0

# Test Dependencies
pytest==7.4.3
pytest-asyncio==0.23.5
pytest-mock==3.12.0
requests==2.31.0
Pillow==10.2.0
imagehash==4.3.1 